SLURM updates, multi-app test suite improvements, and a fix to "make clean"
1. SLURM support for NEMSfv3gfs app's NEMSCompsetRun on uJet and Theia
(see notes below).  Results match the Moab/Torque baselines.

2. Bug fix from Dusan Jovic to eliminate error messages when cleaning
FMS, and remove one temporary file created during the cleaning
process.

3. Major bug fix to the multi-app test system to allow multiple,
multi-app, tests, to happen at the same time.  This bug was causing
the nightly test website to incorrectly report some branch-specific
tests people were doing as the nightly test results.  The change adds
a "test id" that is passed around; the nightly test uses "ngt".
Documentation is updated accordingly.

SLURM SUPPORT NOTES:

1. At this point, only NEMSfv3gfs has been set up for SLURM.  That app
was used because it is the only one that runs on all machines.  Other
apps can be easily ported; the underlying Rocoto and pyprodutil are
the hard parts, and those are done.

2. On Jet, the fv3_wrtGauss_nemsio_c768 hangs with SLURM but not Moab;
all other t/u Jet tests pass on SLURM.  Note that Moab is using tJet,
while SLURM uses uJet, a different set of hardware.  It is possible
this is a uJet issue, not a SLURM issue.  The system logs suggest
there is a firmware bug in the Mellanox Infiniband switches on uJet.
Once xJet is connected to SLURM, we'll have a better idea.  For now,
the failed test is disabled in SLURM.

3. All Theia Moab tests pass on SLURM.  However, on Theia, the SLURM
is misconfigured to think there are only 12 cores per node instead of
24 when task geometries are requested.  I've compensated by telling
the nightly tests that there are only 12 cores per node, which doubles
the number of nodes we use.  To avoid pounding the machine TOO hard,
the Theia SLURM "nightly" tests will only run once a week.  This can
be changed once the admins fix the SLURM misconfiguration.

4. For now, we're putting the GAEA SLURM port on hold.  This is
because GAEA's SLURM configuration may be undergoing a major change in
the near future.  Presently it has a very non-standard configuration
which would require extra effort to support.  The new configuration
may require very different extra effort, and we don't want to do that
twice.